WIP: Settings generator: new version using a Ruby DSL
This replaces the EDD language and associated parser. Instead, the
commands for creating Parameter Framework settings can now be generated
using this Ruby module. Users only need to know a tiny subset of Ruby to
write the configuration files but may also leaverage the full power of
Ruby.

The same concepts and keywords are used in this DSL as in the EDD
language. The C++ domainGeneratorConnector and most of the Python
domainGenerator script will be reusable in order to have the full
pipeline from writing the configuration file to generating the Parameter
Framework Settings XML file.

Signed-off-by: David Wagner <david.wagner@intel.com>